Doc: Fixed web and wasm support.
